requirements.txt
pandas==2.2.2
finbourne-access-sdk >= 0.0.2272, < 2
finbourne-identity-sdk >= 0.0.1845, < 2
lusid-workflow-sdk-preview >= 0.1.587, < 2
lusid-drive-sdk >= 0.1.598, < 2
lusid-jam
lusid-sdk-preview >= 0.11.4713, < 2
dve-lumipy-preview
importlib
luminesce-sdk-preview >= 1.11.140, < 3
lusidtools >= 1.0.0
lusid-scheduler-sdk-preview
fbnconfig==0.1.63